+++ ShapeHandler/Database/Input/ButtonAttributesForm.resx
@@ -117,4 +117,55 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
+  <data name="$this.Icon" type="System.Drawing.Icon, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+    <value>
+        AAABAAEAFCAAAAEAIACoCgAAFgAAACgAAAAUAAAAQAAAAAEAIAAAAAAAAAoAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACOk5MAZHB5HE1eb746TmP+UWZ78YqcrZKouMYWlqi5AAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAALSypwBMXGt2PE5g/ThL
+        Xv9RZXr/fZGp/Juvw6uToa0YdYOMAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAYmtyCUNVZ71OaID/WnSN/3CIof9ugZn/k6nB/4+ktotteX8EAAAAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAHB2dwBsgpU4U2yF6V96lP9keo//e5Ko/3GGnf+Dl6//jaS56neK
+        mDZ4hIwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABLVFkA///sAWmBlJRWcYv/dpCq/7HL
+        3v+NpLj/epCk/2Z4jf+Rpr3/e5Gjk3V9ggMAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAFhf
+        ZAk+TFd5PU9g5DhKXP9GWm3/YXaJ/3OKn/9Za3z/Pk1e/4WbtP99lardfY6aIIOUoAAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAASFJZIjhIVuQ4TmP/Q1px/1Bmf/9bcYv/X3WP/01gdv8zQlH/cISb/36X
+        rPt4jZ5yfYuWA/Pt4wAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABQXGUwOkxd7TpRZ/9GXXX/WnKN/3KJ
+        p/99lLP/do2r/1Zof/9hc4j/f5at/3GJneiFl6U9gZalAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAFFe
+        akpBVmv4Rl53/05mgP9ke5j/gZi3/5KqyP+UrMn/hpu1/2h9lf93jqb/boid/3qQoa+Zp7EKAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAVGNwc01je/9aco7/YHeU/2+GpP+Hnrz/lKvJ/6m/2f+qvtb/hpy1/3KL
+        o/9ogZb/do+i8YeaqD2rpZ0AAAAAAAAAAAAAAAAAAAAAAAAAAABldoSNX3aP/26Fov93jqz/f5e1/2Z/
+        nv+Dm7j/qr/Z/56yx/+JoLb/d5Cm/26InP99lqr/hZqpgKOdlgAAAAAAAAAAAAAAAAAAAAAAAAAAAHeK
+        mpN3kKn/hJ66/5WvzP+ow93/hKC8/4Kcuf+judH/fpeq/3GLnv9wi6D/dpGm/32Wqf+EmqqqnJmUAwAA
+        AAAAAAAAAAAAAAAAAAAAAAAAgZSidoOcs/+Pq8b/oL7Z/7PQ6v+at9P/kKnF/4acsf90i53/do6g/3iR
+        pf94k6b/e5Sm/4edrZu1r6YBAAAAAAAAAAAAAAAAAAAAAAAAAACAkJw/g5mu84ynwf+Pq8b/ka3H/36Y
+        sv90i6L/bYCR/3KEk/92ipn/doub/3SKmv99k6L3i5+uU5SSjQAAAAAAAAAAAAAAAAAAAAAAAAAAAHiE
+        jAtpeIeiZnmM/WR5j/9ieI//X3OH/15xgv9neYr/cIOS/2+Ckf9sfo//b4OU/3yQn7SVoqsOAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAZXN/AF1ocDJQYnPsTWBy/1hqfv9neo3/Z3yQ/2yBlf95j6L/b4SX/2N1
+        iP9neYrpf4+cPYWXpAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD//9cASldiKVFld+kxQlP/QVFg/2V4
+        jf9dcoj/WnGH/22FnP98laz/ZHmN/G+AkHyQmZ4Dzcm+AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        AABOW2ZBSVps9TdHWf82SFf/Q1hs/zZJW/9BVWj/R1lr/3CFmP9qfZDYgI+bHYmVngAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAAAAAADxIUVU9Tl78Okxd/zZIWf8wQE7/JjM9/y89Sf9LXW7/XXGG/2t9
+        jqCrsK4CAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAALzc8SiUwN/kpNT7/NUNP/zlI
+        Vf9EVWT/TWB0/05hdP9XaXv/ZnWCjERQWwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        AABCSUsQNT1DkzA7RPkrNDr/LDY9/0tbav9abH3/VWZ1/EBOXP9BTlnEVl5jFVFXWwAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAS1JWADE1NQg0QEmIMD1H+iMqLf8fJij/JS0v/ztDSKY3QUeOKDQ+/jVF
+        Uvw/SVGcYmZoEUFISwAAAAAAAAAAAAAAAAAAAAAAAAAAAERISgBYXV8IMTpAfjRCTfctOUL/Iyot/yUs
+        Lv8iKCjDKSwsHDY9PzQqNT3vNEVW/zVEUP05QkiZWV5hDExRVAAAAAAAAAAAAAAAAABfY2UAU1RUBDlC
+        SHgyP0r3O0pV/ys1Ov8mLS//LDM01Tc9PzIlKisAO0BADyo0OsIyQ1P/OE5i/yo2QPo0PD+EX2JjBnyB
+        gQAAAAAAS09MAP///wA0PEFaLjpE8DRFVP83Q03/KDAz/ywyNOA1Oz1AAAAAAJaYmwDIpXsAMDk+Zy89
+        Svs2S1//MkVW/yczPPQ3QERc////AAAAAAA6Q0gAOUJHOS04Qd8zRFT/OElY/y85Qf80PEDpO0FEUP//
+        /wBZX2MAAAAAAGVgVgA6QUURKzU9sSk3RP8zRlj/LT5O/y04PsxOVVkbV11gAERLThwtOEHDN0lZ/z5S
+        Zv80Q1D/PEZN61BZXl3CzNMCaG5xAAAAAAAAAAAAAAAAADY+QwAzOj0oKDA0yigzPP8vPUn/KjU++T5G
+        S2llamwGN0FJkTVFVP5DV2r/Oktc/0NPWOlZY2hc0tvgAmtydgAAAAAAAAAAAAAAAAAAAAAAaWhjABgi
+        KAAuNDYxKC8yyycwNf8nMTf/O0RKxj5HTUc0QlDsP1Fj/z1OXP9JVmDlXWZsVu/u6AF3f4QAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAASk9RACAoKwA0OTsrLjQ2si02O/tASlD8MTtDrDZFUv8+TFj/TFhi2lxm
+        bUn///8BgomLAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQkdKACkvMQBGTE8RQ0tPaFNc
+        YJQ2P0bqPklR/U1YX79dZ241BBMjAIuQkwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAen6AAEBHSgAkLDIA////AE9ZX/FXYGWjZGxwHlNeZgCCio4AAAAAAAAAAAAAAAAAAAAAAAAA
+        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAGZrbgB6gYQA/gfwAP4D8AD8AfAA/AHwAPgA
+        8ADwAPAA8ABwAPAAcADwADAA8AAwAPAAMADwABAA8AAQAPAAMADwADAA+ABwAPgAcAD4APAA+ADwAPgB
+        8AD4APAA+ABwAPAAMADgEBAA4DgQAMB4AACAfAAAAP4AAAH/AAAD/4AAD//wAB//8AA=
+</value>
+  </data>
 </root>
