+++ ShapeHandler/Database/Input/InputAttributesForm.resx
