+++ ShapeHandler/Database/InputControlsForm.resx
