+++ ShapeHandler/ShapeDetector.cs
@@ -28,7 +28,7 @@ namespace ShapeHandler
             //this.Application.ActivePage.Shapes[1].AddRow((short)VisSectionIndices.visSectionProp, (short)VisRowIndices.visRowFirst, (short)VisRowTags.visTagDefault);
 
             //ShapeDataForm.Instance.Show();
-            ShapeDataForm shapeDataForm = new ShapeDataForm(shape.ID);
+            InputControlsForm shapeDataForm = new InputControlsForm(shape.ID);
             shapeDataForm.ShowDialog();
         }
